+++ apps/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import "../globals.css";
-import ProviderWrapper from "../exmples/ProviderWrapper";
+import ProviderWrapper from "./exmples/ProviderWrapper";