+++ apps/app/page.tsx
@@ -1,8 +1,5 @@
 'use client'
-import { BadgeField } from '@/components/BadgeField';
-import { Tag } from '@/hooks/useTagSelection';
-import Examples from '../exmples';
-
+import Examples from './exmples';
 
 export default function page() {